\name{MWT-package}
\alias{MWT-package}
\alias{MWT}
\docType{package}
\title{
  mwt: Moderated Welch Test
}
\description{
  Computes FDR based on a moderated version of the Welch test. It's particularly suited for small samples comparisons exprecially with unequal variances
  
}
\details{
  \tabular{ll}{
    Package:\tab mwt \cr
    Type: \tab Package \cr
    Title:\tab Moderated Welch Test \cr
    Version:\tab 0.2.5 \cr
    Date:\tab 2009-10-16 \cr
    License:\tab GPL2 \cr
  }
  ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
  Yudi Pawitan and Stefano Calza
  
  Maintainer: Stefano Calza <stefano.calza@biostatistics.it>
  
}
\references{
  Demissie M, et al: Effects of unequal variance in microarray data analysis 
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
  ~~ Optional links to other man pages, e.g. ~~
  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
